requirements.txt
google-api-core==2.19.1
google-auth==2.30.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-vision==3.7.2
google-cloud-storage==2.17.0
googleapis-common-protos==1.63.2
PyJWT==2.8.0
pillow==10.3.0
fastapi==0.111.0
Werkzeug==3.0.3